Restructures app folder
