+++ src/app/utils/sanitize.test.ts
@@ -1,4 +1,4 @@
-// src/utils/sanitize.test.ts
+// src/app/utils/sanitize.test.ts
 
 import { describe, it, expect } from 'vitest'
 import { sanitizeSentence } from './sanitize'
