+++ src/app/utils/sanitize.ts
@@ -1,4 +1,4 @@
-// src/utils/sanitize.ts
+// src/app/utils/sanitize.ts
 
 export function sanitizeSentence(input: unknown): string | null {
   if (typeof input !== 'string') return null
